Components: Extract Reusable PostTaxonomies components
